Move Bio.Blast.NCBIStandalone and Bio.ParserSupport into Bio.SearchIO._legacy
