+++ Bio/SearchIO/_legacy/__init__.py
@@ -1,0 +1,6 @@
+
+# This file is part of the Biopython distribution and governed by your
+# choice of the "Biopython License Agreement" or the "BSD 3-Clause License".
+# Please see the LICENSE file that should have been included as part of this
+# package.
+"""Legacy functionalities from other parts of Biopython used by SearchIO."""
